Top level CircuitBreakerMetrics object
